fix: moved the logic of highlight walkthrough from containerWidget to widget boundary
